% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustsample.R
\name{rsample}
\alias{rsample}
\title{Call an alias sampler built in Rust}
\usage{
rsample(x, size, replace = TRUE, prob = NULL)
}
\arguments{
\item{x}{either a vector of two or more elements from which to choose, or a positive integer.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement? Defaults to `TRUE`. See details.}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}
}
\description{
Call an alias sampler built in Rust.
}
\details{
This function generally has the same behavior as \code{\link[sample]{base}} with the following exception:

  Only sampling with replacement is supported. If `replace = FALSE`, you will receive an error.
  The `replace` argument is included only for symmetry with arguments in R's base \code{\link[sample]{base}}.
}
