% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustsample.R
\docType{package}
\name{rustsample}
\alias{rustsample}
\title{An example package demonstrating the use of the extendr crate from R.}
\description{
An example package demonstrating the use of the extendr crate from R with an alias sampler.
}
